Switch from template based file structure to category oriented file structure. Update category listing to include article abstract and its tags, and exclude private articles
The previous version relied on templates have the same name as the directories
present in the source directory. This required every document in a directory to
be rendered only by the template set for the directory.

This patch considers each directory as a category and also leverages the front-matter
parsed from the documents to apply the required template, mark it as private etc.